Updates for 64bit mode
Former-commit-id: 891e6c92dd09b5a7680ac22c260fc3c2ccacb383 [formerly a3ac624910264f8609628feb6521a2f006d29ab0]
Former-commit-id: 73e88a98f46a061b51eec9e4270c2f402770efaf

+++ Examples/PSPDFCatalog/Kiosk/PSCImageGridViewCell.m
@@ -98,7 +98,7 @@ static void PSPDFDispatchIfNotOnMainThread(dispatch_block_t block) {
     if (!_progressView.hidden) {
         _progressView.frame = CGRectMake(0.f, 0.f, self.imageView.frame.size.width*0.8, 21.f);
         CGFloat pageLabelHeight = 0.f;//self.isShowingPageLabel ? self.pageLabel.frame.size.width : 0.f;
-        _progressView.center = CGPointMake(roundf(CGRectGetMaxX(self.imageView.frame)/2.f), roundf(CGRectGetMaxY(self.imageView.frame)*9.f/10.f - pageLabelHeight));
+        _progressView.center = CGPointMake(round(CGRectGetMaxX(self.imageView.frame)/2.f), round(CGRectGetMaxY(self.imageView.frame)*9.f/10.f - pageLabelHeight));
     }
 
     self.selectedBackgroundView.frame = CGRectInset(self.imageView.frame, -4.f, -4.f);
